requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
youtube-transcript-api
python-dotenv
tiktoken
deep-translator